Completed 4.22

+++ bloglist/controllers/blog.js
@@ -2,10 +2,7 @@
 /* eslint-disable no-underscore-dangle */
 const blogsRouter = require('express').Router();
 
-const jwt = require('jsonwebtoken');
-
 const Blog = require('../models/blog');
-const User = require('../models/user');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -14,15 +11,7 @@ blogsRouter.get('/', async (request, response) => {
 });
 
 blogsRouter.post('/', async (request, response) => {
-  const { body } = request;
-
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' });
-  }
-
-  const user = await User.findById(decodedToken.id);
+  const { body, user } = request;
 
   const blog = new Blog({
     title: body.title,
@@ -43,15 +32,9 @@ blogsRouter.post('/', async (request, response) => {
 blogsRouter.delete('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id);
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  const { user } = request;
 
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' });
-  }
-
-  const user = await User.findById(decodedToken.id);
-
-  if (blog.user.toString() === user._id.toString()) {
+  if ((blog && user) && blog.user.toString() === user._id.toString()) {
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
   }
